+++ apps/backend/src/modules/users/user.controller.ts
@@ -22,6 +22,14 @@ import { UsersApiPath } from "./libs/enums/enums.js";
  *          email:
  *            type: string
  *            format: email
+ *          name:
+ *            type: string
+ *          createdAt:
+ *            type: string
+ *            format: date-time
+ *          updatedAt:
+ *            type: string
+ *            format: date-time
  */
 class UserController extends BaseController {
 	private userService: UserService;
@@ -53,6 +61,7 @@ class UserController extends BaseController {
 	 *                items:
 	 *                  $ref: "#/components/schemas/User"
 	 */
+
 	private async findAll(): Promise<APIHandlerResponse> {
 		return {
 			payload: await this.userService.findAll(),