+++ apps/frontend/src/libs/components/app/app.tsx
@@ -1,4 +1,5 @@
 import {
+	Header,
 	Loader,
 	Notification,
 	RouterOutlet,
@@ -11,21 +12,19 @@ import {
 	useLocation,
 } from "~/libs/hooks/hooks.js";
 import { actions as userActions } from "~/modules/users/users.js";
+import { QuizForm } from "~/pages/quiz/libs/components/quiz-form/quiz-form.jsx";
 
 const App: React.FC = () => {
 	const { pathname } = useLocation();
 	const dispatch = useAppDispatch();
-	const { dataStatus, user, users } = useAppSelector(({ users }) => ({
+	const { dataStatus } = useAppSelector(({ users }) => ({
 		dataStatus: users.dataStatus,
-		user: users.user,
-		users: users.users,
 	}));
 
 	const isRoot = pathname === AppRoute.ROOT;
 
 	useEffect(() => {
 		if (isRoot) {
-			void dispatch(userActions.loadAll());
 			void dispatch(userActions.getAuthenticatedUser());
 		}
 	}, [isRoot, dispatch]);
@@ -40,13 +39,8 @@ const App: React.FC = () => {
 			{isLoading && <Loader />}
 			{!isLoading && isRoot && (
 				<>
-					<h2>Users: {user?.email}</h2>
-					<h3>Status: {dataStatus}</h3>
-					<ul>
-						{users.map((user) => (
-							<li key={user.id}>{user.email}</li>
-						))}
-					</ul>
+					<Header />
+					<QuizForm />
 				</>
 			)}
 			<Notification />