+++ apps/frontend/src/modules/users/slices/actions.ts
@@ -1,25 +1,12 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
-import {
-	type UserGetAllItemResponseDto,
-	type UserGetAllResponseDto,
-} from "~/modules/users/users.js";
+import { type UserDto } from "~/modules/users/users.js";
 
 import { name as sliceName } from "./users.slice.js";
 
-const loadAll = createAsyncThunk<
-	UserGetAllResponseDto,
-	undefined,
-	AsyncThunkConfig
->(`${sliceName}/load-all`, (_, { extra }) => {
-	const { userApi } = extra;
-
-	return userApi.getAll();
-});
-
 const getAuthenticatedUser = createAsyncThunk<
-	UserGetAllItemResponseDto,
+	UserDto,
 	undefined,
 	AsyncThunkConfig
 >(`${sliceName}/get-authenticated-user`, (_, { extra }) => {
@@ -28,4 +15,4 @@ const getAuthenticatedUser = createAsyncThunk<
 	return userApi.getAuthenticatedUser();
 });
 
-export { getAuthenticatedUser, loadAll };
+export { getAuthenticatedUser };