+++ apps/mobile/src/libs/components/components.tsx
@@ -9,6 +9,7 @@ export { LoaderWrapper } from "./loader-wrapper/loader-wrapper";
 export { Planet } from "./planet/planet";
 export { ScreenWrapper } from "./screen-wrapper/screen-wrapper";
 export { Text } from "./text/text";
+export { ToastMessage } from "./toast-message/toast-message";
 export { default as MaskedView } from "@react-native-masked-view/masked-view";
 export {
 	ActivityIndicator,