+++ apps/mobile/src/libs/helpers/get-is-ios/get-is-ios.ts
@@ -1,4 +1,0 @@
-import { Platform } from "~/libs/components/components";
-
-const getIsIos = () => Platform.OS === "ios";
-export { getIsIos };